+++ tests/generator_clone.rs
@@ -32,9 +32,9 @@ fn gen_iter_return_clone() {
     let mut g_cloned = g.clone();
     assert_eq!((&mut g_cloned).next(), Some(2));
     assert_eq!((&mut g_cloned).next(), None);
-    assert_eq!(g_cloned.return_or_self().ok(), Some("done"));
+    assert_eq!(g_cloned.try_get_return().ok(), Some("done"));
 
     assert_eq!((&mut g).next(), Some(2));
     assert_eq!((&mut g).next(), None);
-    assert_eq!(g.return_or_self().ok(), Some("done"));
+    assert_eq!(g.try_get_return().ok(), Some("done"));
 }